ci(wasm): enforce full wasm build, vet, and test compilation
The wasm build guard only compiled the root package, so wasm-tagged code
outside it rotted unnoticed: the js/wasm build of ./... was broken
(auth GetProfile missing on the WASM config manager) and go vet under the
wasm tags failed to compile three native-only test files.

- add GetProfile to the WASM ConfigManager (and its interface), mirroring
  native, so auth status compiles and works in the browser build
- gate the native-only test files (messages_test, changes_test,
  login_test) with build tags matching their sibling native tests
- fix the wasm-tagged lint findings (errcheck, unparam, staticcheck) and
  move two native-only output helpers into common_native.go, so wasm lints
  under the same golangci-lint ruleset with no linters disabled
- Makefile: wasm-build-guard now builds ./...; add a wasm-vet target
- wasm.yml: run wasm vet in CI (compiles every wasm test package)
- consolidate lint on one pinned golangci-lint version (drop the
  duplicate v2.11.4 job in test.yml) and add a golangci-wasm job that
  lints wasm-tagged files

ESD-1578
