+++ .github/workflows/lint.yml
@@ -24,3 +24,23 @@ jobs:
         with:
           version: 'v2.12.2'  # Pin the golangci-lint binary version to avoid CI drift
           args: --timeout=10m
+
+  golangci-wasm:
+    name: lint (js/wasm)
+    runs-on: ubuntu-latest
+    # Lint the wasm-tagged files with the same ruleset. golangci-lint only sees
+    # one GOOS/GOARCH per run, so the default job above never covers them.
+    env:
+      GOOS: js
+      GOARCH: wasm
+    steps:
+      - uses: actions/checkout@93cb6efe18208431cddfb8368fd83d5badbf9bfd  # v5.0.1
+      - uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16  # v6.5.0
+        with:
+          go-version-file: 'go.mod'
+          cache: false
+      - name: Run golangci-lint
+        uses: golangci/golangci-lint-action@82606bf257cbaff209d206a39f5134f0cfbfd2ee  # v9.2.1
+        with:
+          version: 'v2.12.2'  # Pin the golangci-lint binary version to avoid CI drift
+          args: --build-tags js,wasm --timeout=10m
