+++ .github/workflows/test.yml
@@ -26,19 +26,6 @@ jobs:
           files: coverage.out
           fail_ci_if_error: false
 
-  lint:
-    name: Lint
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@93cb6efe18208431cddfb8368fd83d5badbf9bfd  # v5.0.1
-      - uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16  # v6.5.0
-        with:
-          go-version-file: 'go.mod'
-      - name: Install golangci-lint
-        run: go install github.com/golangci/golangci-lint/v2/cmd/golangci-lint@v2.11.4
-      - name: Run golangci-lint
-        run: golangci-lint run --timeout=5m
-
   e2e-hermetic:
     name: E2E Tests (Hermetic)
     # Run on Unix and Windows so OS-specific regressions in the cross-platform
