+++ internal/base/output/messages_wasm.go
@@ -209,7 +209,7 @@ func PrintSuccess(format string, noColor bool, args ...interface{}) {
 	} else {
 		output = color.GreenString("✓ ") + msg + "\n"
 	}
-	wasm.WasmOutputBuffer.Write([]byte(output))
+	_, _ = wasm.WasmOutputBuffer.Write([]byte(output))
 }
 
 // PrintError overrides the base function for WASM to capture output. In json
@@ -227,7 +227,7 @@ func PrintError(format string, noColor bool, args ...interface{}) {
 	} else {
 		output = color.RedString("✗ ") + msg + "\n"
 	}
-	wasm.WasmOutputBuffer.Write([]byte(output))
+	_, _ = wasm.WasmOutputBuffer.Write([]byte(output))
 	markErrorEmitted()
 }
 
@@ -243,7 +243,7 @@ func PrintWarning(format string, noColor bool, args ...interface{}) {
 	} else {
 		output = color.YellowString("⚠ ") + msg + "\n"
 	}
-	wasm.WasmOutputBuffer.Write([]byte(output))
+	_, _ = wasm.WasmOutputBuffer.Write([]byte(output))
 }
 
 // PrintInfo overrides the base function for WASM to capture output
@@ -258,7 +258,7 @@ func PrintInfo(format string, noColor bool, args ...interface{}) {
 	} else {
 		output = color.BlueString("ℹ ") + msg + "\n"
 	}
-	wasm.WasmOutputBuffer.Write([]byte(output))
+	_, _ = wasm.WasmOutputBuffer.Write([]byte(output))
 }
 
 // PrintPlain overrides the base function for WASM to capture output.
@@ -267,7 +267,7 @@ func PrintPlain(format string, _ bool, args ...interface{}) {
 		return
 	}
 	msg := fmt.Sprintf(format, args...)
-	wasm.WasmOutputBuffer.Write([]byte(msg + "\n"))
+	_, _ = wasm.WasmOutputBuffer.Write([]byte(msg + "\n"))
 }
 
 // PrintNewline writes a blank line to the WASM output buffer, suppressed in quiet mode.
@@ -275,7 +275,7 @@ func PrintNewline() {
 	if IsQuiet() {
 		return
 	}
-	wasm.WasmOutputBuffer.Write([]byte("\n"))
+	_, _ = wasm.WasmOutputBuffer.Write([]byte("\n"))
 }
 
 // ClearScreen is a no-op in the WASM environment.
