+++ internal/commands/config/login_test.go
@@ -1,3 +1,5 @@
+//go:build !js && !wasm
+
 package config
 
 import (
